+++ src/talk_tcpr.cpp
@@ -10,7 +10,7 @@ std::string w_tcpr(std::string vertical_align, std::string text_direction,
                    bool do_bgimg, std::string bgimg_rid, std::string bgimg_path,
                    IntegerVector btlr_red, IntegerVector btlr_green,
                    IntegerVector btlr_blue, IntegerVector btlr_alpha,
-                   CharacterVector type, IntegerVector width ) {
+                   CharacterVector type, DoubleVector width ) {
   tcpr tcpr_(vertical_align, text_direction,
              mb, mt, ml, mr,
              shd_r, shd_g, shd_b, shd_a,
@@ -27,7 +27,7 @@ std::string a_tcpr(std::string vertical_align,std::string text_direction,
                    bool do_bgimg, std::string bgimg_rid, std::string bgimg_path,
                    IntegerVector btlr_red, IntegerVector btlr_green,
                    IntegerVector btlr_blue, IntegerVector btlr_alpha,
-                   CharacterVector type, IntegerVector width) {
+                   CharacterVector type, DoubleVector width) {
   tcpr tcpr_(vertical_align, text_direction,
              mb, mt, ml, mr,
              shd_r, shd_g, shd_b, shd_a,
@@ -44,7 +44,7 @@ std::string css_tcpr(std::string vertical_align,std::string text_direction,
                    bool do_bgimg, std::string bgimg_rid, std::string bgimg_path,
                    IntegerVector btlr_red, IntegerVector btlr_green,
                    IntegerVector btlr_blue, IntegerVector btlr_alpha,
-                   CharacterVector type, IntegerVector width ) {
+                   CharacterVector type, DoubleVector width ) {
   tcpr tcpr_(vertical_align, text_direction,
              mb, mt, ml, mr,
              shd_r, shd_g, shd_b, shd_a,
